Use AWS::Core::Data
When reducing a hash thats of the type 'AWS::Core::Data', there's a bug
which means the value of the hash isn't available in the reduce block.

To get around this, we call `to_h` on the `AWS::Core::Data` which
returns us a raw hash that can correctly be looped over.