Remove dalli-elasticache wrapper
The dalli-elasticache wrapper is for auto-discovering elasticache endpoints.
However due to recent updates with AWS Elasticache, this behaviour is no longer
needed. What's more is this gem has a non parameterised 60s timeout when
autodiscovering and causes Alephant to hang for a long time on each request
if the provided elasticache endpoint is unreachable.

This change switches to using the Dalli client directly.